provide a sync api
This is an attempt to provide a sync api, which makes much more sense
for expect. The idea is to run the event loop in a separate thread so
that we can call `Future::wait()` in the client handle without
preventing the event loop from making progress.

Note that this currently rely on
https://github.com/rust-lang/rust/pull/47760

Fwiw, running our own even loop is not considered good practice, but it
the approach reqwest has taken too, so I guess it's not that bad.